move managed_state to v2
