+++ src/stream/src/executor_v2/managed_state/aggregation/extreme_serializer.rs
