+++ src/stream/src/executor_v2/managed_state/top_n/top_n_state.rs
@@ -24,9 +24,9 @@ use risingwave_storage::cell_based_row_deserializer::CellBasedRowDeserializer;
 use risingwave_storage::storage_value::StorageValue;
 use risingwave_storage::{Keyspace, StateStore};
 
-use crate::executor::managed_state::flush_status::BtreeMapFlushStatus as FlushStatus;
-use crate::executor::managed_state::top_n::variants::*;
-use crate::executor::managed_state::top_n::PkAndRowIterator;
+use super::super::flush_status::BtreeMapFlushStatus as FlushStatus;
+use super::variants::*;
+use super::PkAndRowIterator;
 
 /// This state is used for several ranges (e.g `[0, offset)`, `[offset+limit, +inf)` of elements in
 /// the `AppendOnlyTopNExecutor` and `TopNExecutor`. For these ranges, we only care about one of the
@@ -386,9 +386,8 @@ mod tests {
     use risingwave_storage::memory::MemoryStateStore;
     use risingwave_storage::{Keyspace, StateStore};
 
+    use super::super::variants::TOP_N_MAX;
     use super::*;
-    use crate::executor::managed_state::top_n::top_n_state::ManagedTopNState;
-    use crate::executor::managed_state::top_n::variants::TOP_N_MAX;
     use crate::row_nonnull;
 
     fn create_managed_top_n_state<S: StateStore, const TOP_N_TYPE: usize>(
